more minor progression
artifacts display (except artifact id for some reason) and when you click on the artifact, it loads it into the list.
Need to figure out how to get to the artifact details page.
warehouses list is looking better. still cannot get it to go to the warehouse details page which should mimic the artifact list except show only items from the chosen warehouse.